[FIX] GH-21 Initial working migration to tsdx
