+++ .gitignore
@@ -2,15 +2,12 @@
 
 # dependencies
 /node_modules
-/.pnp
-.pnp.js
-/packages
 
 # testing
 /coverage
 
 # production
-/build
+/dist
 
 # misc
 .DS_Store
@@ -19,7 +16,4 @@
 .env.test.local
 .env.production.local
 .eslintcache
-
-npm-debug.log*
-yarn-debug.log*
-yarn-error.log*
+*.log
