+++ .storybook/preview.js
@@ -1,10 +1,5 @@
-
+// https://storybook.js.org/docs/react/writing-stories/parameters#global-parameters
 export const parameters = {
-  actions: { argTypesRegex: "^on[A-Z].*" },
-  controls: {
-    matchers: {
-      color: /(background|color)$/i,
-      date: /Date$/,
-    },
-  },
-}
+  // https://storybook.js.org/docs/react/essentials/actions#automatically-matching-args
+  actions: { argTypesRegex: '^on.*' },
+};
