+++ .vscode/settings.json
@@ -1,3 +1,14 @@
 {
-  "editor.defaultFormatter": "esbenp.prettier-vscode"
+  "editor.formatOnPaste": false,
+  "editor.formatOnSave": true,
+  "files.eol": "\n",
+  "editor.tabSize": 2,
+  "eslint.validate": [
+    "javascript",
+    "javascriptreact",
+    "typescript",
+    "typescriptreact"
+  ],
+  "editor.defaultFormatter": "esbenp.prettier-vscode",
+  "prettier.endOfLine": "lf"
 }
