+++ stories/GaugeChart.stories.tsx
@@ -1,17 +1,17 @@
 import React from 'react';
 import { Story, Meta } from '@storybook/react';
 
-import { RadarChart } from './RadarChart';
+import { GaugeChart } from '../src/charts/GaugeChart';
 
 export default {
-  title: 'RadarChart',
-  component: RadarChart,
+  title: 'GaugeChart',
+  component: GaugeChart,
   argTypes: {
     backgroundColor: { control: 'color' },
   },
 } as Meta;
 
-const Template: Story = (args) => <RadarChart {...args} />;
+const Template: Story = (args) => <GaugeChart {...args} />;
 
 export const Primary = Template.bind({});
 Primary.args = {};
