+++ stories/RadarChart.stories.tsx
@@ -1,17 +1,17 @@
 import React from 'react';
 import { Story, Meta } from '@storybook/react';
 
-import { GaugeChart } from './GaugeChart';
+import { RadarChart } from '../src/charts/RadarChart';
 
 export default {
-  title: 'GaugeChart',
-  component: GaugeChart,
+  title: 'RadarChart',
+  component: RadarChart,
   argTypes: {
     backgroundColor: { control: 'color' },
   },
 } as Meta;
 
-const Template: Story = (args) => <GaugeChart {...args} />;
+const Template: Story = (args) => <RadarChart {...args} />;
 
 export const Primary = Template.bind({});
 Primary.args = {};
